+++ server/graphql/resolvers/project/ProjectResolver.ts
@@ -1,3 +1,4 @@
+import _ from 'lodash'
 import 'reflect-metadata'
 import { Arg, Authorized, Ctx, Mutation, Query, Resolver } from 'type-graphql'
 import { Service } from 'typedi'
@@ -12,7 +13,6 @@ import {
   ProjectInput,
   ProjectOptions
 } from '../types'
-import _ from 'lodash'
 
 /**
  * Resolver for `Project`.
